requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
beautifulsoup4==4.12.2
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
idna==3.6
    # via requests
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
soupsieve==2.5
    # via beautifulsoup4
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
urllib3==2.1.0
    # via requests
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
